Use 'new' with sdk/net/xhr XMLHttpRequest
I couldn't get this addon to work so I went into about:addons and ran the debugger and saw some error saying that the `XMLHttpRequest` object from `sdk/xhr` needs to be instantiated with `new`. I edited the source of the extension directly in my profile folder and restarted the browser - lo and behold, it works!